Update For Puzzle Game
Add max distance formula to avoid from unsolvable puzzles. For this ı handle all piece of tiles and take sqrt (x^2+y^2+z^2) based on this value we have a limit for shuffling. The max distance mean a tile which has not an adjencincy of empty tile. After that I reimplement the shuffling methode to before every shuffle handle that proceed shuffle if the selected tile with empthy tile have distanceless then max distance .